feat(org): show img that's used for id, elisp link
Below is an explanation of how the image-url in "mailto:" link
description can be used to displayed as an inline image
link (585ffc18e2ce):

Sometimes you want to use a mail image as a link. This is easy to do
with the http: protocol, but not with the mailto: protocol (which is
used for email links) because the mailto: protocol does not have a way
to embed an image in, and it is not possible to embed an image in a
mailto: link.

However, you can embed an image in a mailto: link in the link
description by setting the :image-data-fun property to a function like
`+org-http-with-desc-image-data-fn' that takes the image-url in
description if it's available and returns the image data.

Actually, what I want to do is different from org-yt's
`org-display-user-inline-images' function which is designed to scan for
image data links only from the link part. The org-yt use case is to use
the :image-data-fun property to embed the image in the link (which is
used for the YouTube link) automatically like a video thumbnail. So it's
no need to use the description as the image data. Therefore, my
intention to use description as the image data is not working as I
expected for the mailto: protocol even if I put the image data link in the
description because it only checks the link (which is the mailto:
protocol) and this link type does not have a way to embed an image in,
and it is not possible to embed an image in a mailto: link. That's why
mailto: doesn't have a :image-data-fun parameter by default.

Signed-off-by: Nopanun Loachunhanun <nopanun@pm.me>