tweak(org): set mailto :image-data-fun
Sometimes you want to use a mail image as a link. This is easy to do with the http:
protocol, but not with the mailto: protocol (which is used for email links)
because the mailto: protocol does not have a way to embed an image in, and it is
not possible to embed an image in a mailto: link.

However, you can embed an image in a mailto: link in the link description by
setting the :image-data-fun property to a function like
`+org-http-with-desc-image-data-fn' that takes the description and returns the
image data.

Actually, what I tried to do was the difference use case from org-yt's
`org-display-user-inline-image' function was designed to scan the links (not
description). The org-yt use case is to use the :image-data-fun property to embed
the image in the link (which is used for the YouTube link) automatically like a VDO
thumbnail. So it's no need to use the description as the image data. Therefore, my
intention to use description as the image data is not working as I expected for
the mailto: protocol even I put the image data link in the description because it
only checks the link (which is the mailto: protocol) and this link type does not
have a way to embed an image in, and it is not possible to embed an image
in a mailto: link. That's why mailto doesn't have a :image-data-fun parameter by
default.