fix: void function error after entry to minibuffer
Debugger entered--Lisp error: (void-function `minibuffer-header-mode')
or Symbol’s function definition is void: `minibuffer-header-mode'.

This means that the function is not defined or not visible to Emacs yet due
to it is still not being loaded by Doom Emacs because of its lazy loading.
(Doom lazy loads packages to boost Emacs startup speed.)

Normally, this is not a problem because Doom can know exactly what the
function is, without requiring the package that defines it to be loaded first
with the needs of an “autoload cookie” (a magic autoload comment consists of
‘;;;###autoload’, on a line by itself, just before the real definition of the
function in its autoloadable source file). However, it is not the case for
`minibuffer-header-mode' because it is defined in the package
‘minibuffer-header’, which does not have any lines of autoload cookies.

Fix: #1
Ref: rougier/minibuffer-header@fe5d90d3f13c